// frontend/rodia-monorepo/apps/mobile/src/shared/api/generated/schemas/shipperPaymentMethodSummaryResponse.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface ShipperPaymentMethodSummaryResponse {
  id?: string;
  type?: string;
  provider?: string;
  holderName?: string;
  last4?: string;
  registeredAt?: string;
  lastUsedAt?: string;
  usageCount?: number;
  isDefault?: boolean;
}